splitting the controller and service
