+++ movie-info-service/src/main/java/com/example/movieinfoservice/controller/MovieController.java
@@ -1,0 +1,26 @@
+package com.example.movieinfoservice.controller;
+
+import com.example.movieinfoservice.entity.Movie;
+import com.example.movieinfoservice.service.MovieService;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/movies")
+public class MovieController {
+
+    @Autowired
+    private MovieService movieService;
+
+    public MovieController(MovieService movieService) {
+        this.movieService = movieService;
+    }
+
+    @RequestMapping("/{movieId}")
+    public Movie getMovieInfo(@PathVariable("movieId") String movieId) {
+        return movieService.getMovieInfo(movieId);
+    }
+}
