+++ movie-info-service/src/main/java/com/example/movieinfoservice/repository/MovieRepository.java
@@ -1,0 +1,9 @@
+package com.example.movieinfoservice.repository;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import com.example.movieinfoservice.entity.Movie;
+
+public interface MovieRepository extends MongoRepository<Movie, String> {
+
+}
