--- movie-info-service/src/main/java/com/example/movieinfoservice/resources/MovieResource.java
@@ -1,33 +1,0 @@
-package com.example.movieinfoservice.resources;
-
-import com.example.movieinfoservice.models.Movie;
-import com.example.movieinfoservice.models.MovieSummary;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.client.RestTemplate;
-
-@RestController
-@RequestMapping("/movies")
-public class MovieResource {
-
-    @Value("${api.key}")
-    private String apiKey;
-
-    private RestTemplate restTemplate;
-
-    public MovieResource(RestTemplate restTemplate) {
-        this.restTemplate = restTemplate;
-    }
-
-    @RequestMapping("/{movieId}")
-    public Movie getMovieInfo(@PathVariable("movieId") String movieId) {
-        // Get the movie info from TMDB
-        final String url = "https://api.themoviedb.org/3/movie/" + movieId + "?api_key=" + apiKey;
-        MovieSummary movieSummary = restTemplate.getForObject(url, MovieSummary.class);
-
-        return new Movie(movieId, movieSummary.getTitle(), movieSummary.getOverview());
-    }
-}
