+++ movie-info-service/src/main/resources/application.properties
@@ -2,3 +2,4 @@ spring.application.name=movie-info-service
 server.port=8082
 
 api.key=b0af84324ea41ca39f4311c9156e94e3
+spring.data.mongodb.uri=mongodb://localhost:27017/moviedb
